requirement.txt
Package    Version
---------- -------
asgiref    3.3.1  
cffi       1.14.5 
colormath  3.0.0  
decorator  4.4.2  
Django     3.1.7  
future     0.18.2 
kociemba   1.2.1  
networkx   2.5.1  
numpy      1.20.2 
Pillow     8.2.0  
pip        19.0.3 
pycparser  2.20   
pytz       2021.1 
scipy      1.6.2  
setuptools 40.8.0 
sqlparse   0.4.1  
